Separate code hook for api calls
